chore: Refactor buildTransaction for VRF/RR
